requirements.txt
tkinter 
Pillow
numpy
pandas
tensorflow
keras